DEV: Convert activation modals to component-based API
