--- app/assets/javascripts/discourse/app/templates/modal/activation-resent.hbs
@@ -1,7 +1,0 @@
-<DModalBody>
-  {{html-safe
-    (i18n "login.sent_activation_email_again" currentEmail=this.currentEmail)
-  }}
-</DModalBody>
-
-<ModalFooterClose @closeModal={{route-action "closeModal"}} />
